+++ tests/unit/algo/classification/test_huggingface_model.py
@@ -1,6 +1,8 @@
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
+from unittest.mock import MagicMock
+
 import pytest
 import torch
 from otx.core.data.entity.base import OTXBatchLossEntity
@@ -9,6 +11,7 @@ from otx.core.exporter.native import OTXNativeModelExporter
 
 SKIP_TRANSFORMERS_TEST = False
 try:
+    from otx.algo.classification import huggingface_model as target_file
     from otx.algo.classification.huggingface_model import HuggingFaceModelForMulticlassCls
     from transformers.modeling_outputs import ImageClassifierOutput
 except ImportError:
@@ -67,3 +70,23 @@ class TestHuggingFaceModelForMulticlassCls:
             output = fxt_multi_class_cls_model.forward_for_tracing(
                 image=torch.randn(1, 3, 224, 224),
             )
+
+    @pytest.fixture()
+    def mock_pretrainedconfig(self, mocker) -> MagicMock:
+        mock_obj = mocker.patch.object(target_file, "PretrainedConfig")
+        mock_obj.get_config_dict.return_value = ({"image_size": 224}, None)
+        return mock_obj
+
+    @pytest.fixture()
+    def mock_automodel(self, mocker) -> MagicMock:
+        return mocker.patch.object(target_file, "AutoModelForImageClassification")
+
+    def test_set_input_size(self, mock_pretrainedconfig, mock_automodel):
+        input_size = (1, 3, 300, 300)
+        HuggingFaceModelForMulticlassCls(
+            model_name_or_path="facebook/deit-tiny-patch16-224",
+            label_info=10,
+            input_size=input_size,
+        )
+
+        assert mock_automodel.from_pretrained.call_args.kwargs["image_size"] == input_size[-1]